LayoutTests/media/video-playback-interrupted-expected.txt


RUN(internals.setMediaSessionRestrictions("videoaudio", "InterruptedPlaybackNotPermitted"))
RUN(video.src = findMediaFile("video", "content/test"))
EXPECTED (video.paused == 'true') OK
RUN(internals.beginMediaSessionInterruption("System"))
RUN(video.play().then(didResolvePromise).catch(didRejectPromise))
RUN(internals.endMediaSessionInterruption("MayResumePlaying"))
Promise resolved. OK
EXPECTED (video.paused == 'false') OK
END OF TEST
